refactor(client)!: move record size estimator classes under org.apache.hudi.execution.estimator
Relocate RecordSizeEstimator (@PublicAPIClass), AverageRecordSizeEstimator and
RecordSizeEstimatorFactory from the 3-class top-level org.apache.hudi.estimator
package to org.apache.hudi.execution.estimator, beside the write-execution
machinery that consumes them. Config docs for
hoodie.record.size.estimator.class updated to the new FQNs.

BREAKING CHANGE: users who explicitly set hoodie.record.size.estimator.class
to org.apache.hudi.estimator.AverageRecordSizeEstimator get a loud
ClassNotFoundException at write start (fix: update the config value); custom
estimators compiled against the old package fail with NoClassDefFoundError
(fix: import change + recompile). Config default tracks the move
automatically; no compat stub by design - every failure is loud, no silent
misbehavior is possible. Accepted by VC 2026-07-05, tracked as BC-5.
