+++ hudi-client/hudi-client-common/src/test/java/org/apache/hudi/execution/estimator/TestAverageRecordSizeEstimator.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.estimator;
+package org.apache.hudi.execution.estimator;
 
 import org.apache.hudi.common.fs.FSUtils;
 import org.apache.hudi.common.model.HoodieCommitMetadata;
